+++ Vectorizing Animation Video/math/Centroid.h
@@ -12,16 +12,16 @@ public:
 		int psize = (int)poly.const_Points().size();
 		for (int i = 0; i < psize; i++) 
 			m_centroid += poly.const_Points()[i];
-		m_centroid /= (float)psize;
+		m_centroid /= (double)psize;
 		m_radius = 0;
 		for (int i = 0; i < psize; i++) 
 		{		
-			float tmp = m_centroid.dotProduct(poly.const_Points()[i]);
+			double tmp = m_centroid.dotProduct(poly.const_Points()[i]);
 			if (tmp>m_radius)
 				m_radius = tmp;
 		}
 	}
 	Vector2 m_centroid;
-	float	m_radius;
+	double	m_radius;
 };
